Make all dependencies in the app module to be created by Dagger in a Service Locator way
One step more, but not still what we want. The way that the main module is configured is a little bit strange, why we need to have two @Provides methods (One for Activity other for Context)? In the next step we need to break the main module to separate some responsibilities