Break main module to separate the dependencies to their respective scope
In this commit we have very clear what dependencies are in Activity scope (MainModule) and what dependencies are in the Application scope (DatabaseModule and ServiceModule). In this way our Dagger setup got a little better. In the next step lets use construction injection to remove some dependency creation code in the modules