Uses constructor injection via @Inject to remove dependency creation in modules
In this commit we used Dagger constructor injection to setup our dependencies. Notice that the classes created on the modules are interfaces or classes that we dont have control, all other classes should be injected by constructor injection via @Inject. Also notice that the @Inject annotation worked across different modules.