Uses field injection to inject presenter in activity
Until now we were creating our presenter in a service locator way... We changed that to use the field injection feature from Dagger. The main problem now is that we have all dependencies in Dagger graph being recreated on every configuration change (Screen rotation for instance). So in the next commit we would have to separate the application scope from the activity scope using subcomponents. By doing that, dependencies like Retrofit or the database dont have to be recreated every time