Break component in ApplicationComponent (Application Scoped) and MainComponent (Activity Scoped)
In this commit we created a custom Application class that holds our ApplicationComponent, in this way we dont need to call the application component builder.build() in every configuration change. By using the subcomponent feature of Dagger, we can use the dependencies from both components at the same time with very clear separation of concerns. This is a great step forward but we are not done yet! On Every configuration change, the dependencies from de ApplicationModule are still being recreated. Lets use the built-in scoped anottations to solve this problem on the next commit