Annotate heavy dependencies with @Singleton to avoid recreation on configuration changes
We used @Singleton in this commit to avoid heavy class recreation on configuration changes. Internally, Dagger do some operations to check if the class was previously created or not, for that reason we should use @Singleton with caution. It is recommended to use it only with classes that contains state or are very heavy to create/initialize (Database for instance). Notice that not only the methods in the modules were annotated with @Singleton but also the associated component as well