Creates custom builder on ApplicationComponent to hold context reference
By using @Component.Builder annotation we can create custom builders to our components. We normally use that when our modules are holding state... So other way to do that is to delegate the state holding to the component and not to the module. Under the hoods this change can improve the performance, specially if your component holds many modules with state