Changes injection to alternative solution that uses two components instead of component+subcomponent
In this solution we use one component to be a dependency of the other component. For us to do that we had to do some steps. First of all we had to say that the MainComponent depends on AplicationComponent via the dependencies parameter in @Component annotation, after that we had to expose the dependencies that the MainComponent needed in the ApplicationComponent (UserDAO and MainService). At least we had to create a custom builder and custom scope in MainComponent in order to comunicate properly with the ApplicationComponent