Setup instrumented tests to use retrofit service that points to mocked endpoint
In this commit we had to configure dagger in the instrumented test package. As part of the configuration we had to create a new Application class in the tests and override the main application by using the instrumentation test runner in the build gradle file. This test application uses a test version of the dependencies. In order to make the application usable and visible for the feature two module, we had to use an interface to achieve a good communication between them